add support for yeelights get_prop command
get_prop takes a string as payload to specify the property. To be able
to read the result data the node will now output a message with the
payload set to the result of the command executed.

This allows the user to access data such as the current power status.