Deconstruct array input to allow passing duration
https://lsong.org/node-yeelight/Yeelight.html#set_bright

I'd like to set a longer duration when changing the brightness in my scheduled automations. This would allow people to pass in an array like `[1,"smooth",10000]` to change brightness to 1% over 10 seconds. It would keep working normally for people who only pass a single value. The docs might also need updating to reflect this option.